Split codebase into three sections
There are three specific parts to the project:

Library module:
The OpsLogger class and LogMessage interface. This is all
of the production code and its direct tests. Production code
that performs logging will only depend on this module.

Support module:
Support classes that help test code that *uses* OpLogger,
the contract test for enums that implement LogMessage,
and the tool that generates automatic documentation from
LogMessage enums. Application unit tests and build
infrastructure will depend on this module.

Sample-usage module:
Serves as a guide of how you would use this library from
an application, and what kinds of tests you would write.
This module is basically documentation. Nobody would
ever depend on this module directly, but they might look
at the source code and copy bits of it.