Add checks for too many arguments to a mock OpsLogger
String.format checks "for not enough arguments provided",
but not for "too many arguments provided". Check this in
our code by removing the last argument, and attempting to
format the string. If this doesn't cause an error, then the final
argument isn't used, and too many arguments have been
provided.