Add a sample spring config to demonstrate usage
Add a dependency on spring and a sample application context
to the sample-usage project.

Modify OpsLogger factory to use a property (setDestination)
instead of a method (loggingTo) — the spring example is
much more straightforward with this change (and therefore
real spring configuration will be as well). Spring doesn't seem
to mind that the "property" setter returns an instance (even
though that technically makes it not a property), so both
property support and method chaining is possible.