Add StackTraceProcessor and SimpleStackTraceProcessor
StackTraceProcessor processes Throwable instances for
inclusion in log messages — with the intent being that most
implementations will somehow condense the stack trace
output to a single line.

SimpleStackTraceProcessor is basically an identity transform
that preserves the entire multi-line stack trace in the log
message.

Extract TestPrintStream to a top-level test class as it is used
by more than one test.