OpsLoggerTestDouble instances can't be closed
OpsLogger instances should almost never be closed by the
classes that log messages. Instead, the instances (which
should be singletons) should be closed by the IoC container
(aka Spring, Guice, etc) when the object graph is closed down.

OpsLoggerTestDouble throws an IllegalStateException when
close() is called to encourage the correct behaviour in app
code that performs logging.