Timestamps should include milliseconds on a whole second
Use a custom DateTimeFormatter to override the default
behaviour, which will use zero, three, six, or nine decimal
places depending on the nanosecond count. In practice, this
means that an instance will have zero decimal places if it is
exactly on an even second.

Always providing milliseconds keeps the output lined up.