OpsLoggerFactory is more symlink-friendly
OpsLoggerFactory will not create parent directories for a log
file if the immediate parent of the logfile is a symlink.

Similarly, OpsLoggerFactory will not create directories for
a stack trace storage path if the path is a symlink.

Creating directories for an existing symlink doesn't work
(it fails with a FileAlreadyExists exception, because symlinks
are *files* that link to directories), so we just assume that their
targets exist already.