Correlation IDs are provided as a map and rendered with names
Feedback from a client is that rendering with names is more useful than anonymous
rendering as a list. While the representation is larger, it's also much more useful when
it's not immediately obvious what type the correlator has by it's value (when it is an
integer, for example).

This approach is also much more useful when there are a disjoint set of correltors that
are highly optional (for example, regular users have a user id, admin users have a user
name, and department, and a user they are operating on behalf of).