Fix mutability check
Allow immutable objects to be passed as log message format
string arguments, as well as just "effectively immutable"
objects (which is actually a weaker form of immutability). The
check "areEffectivelyImmutable()" checks only that arguments
are effectively immutable, not effectively immutable or better.

Improve the check to allow objects (like java.util.UUID) which
are entirely immutable as well.

Resolves #5.